Remove unnecessary boxing

+++ core/src/main/java/org/trimou/util/Checker.java
@@ -30,17 +30,17 @@ import org.trimou.annotations.Internal;
 @Internal
 public final class Checker {
 
-    private static final Long LONG_ZERO = Long.valueOf(0L);
+    private static final Long LONG_ZERO = 0L;
 
-    private static final Integer INTEGER_ZERO = Integer.valueOf(0);
+    private static final Integer INTEGER_ZERO = 0;
 
-    private static final Short SHORT_ZERO = Short.valueOf((short) 0);
+    private static final Short SHORT_ZERO = 0;
 
-    private static final Byte BYTE_ZERO = Byte.valueOf((byte) 0);
+    private static final Byte BYTE_ZERO = 0;
 
-    private static final Double DOUBLE_ZERO = Double.valueOf(0.0d);
+    private static final Double DOUBLE_ZERO = 0.0;
 
-    private static final Float FLOAT_ZERO = Float.valueOf(0.0f);
+    private static final Float FLOAT_ZERO = 0.0F;
 
     private Checker() {
     }
